color-pallete-generator/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
imageio==2.28.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
MarkupSafe==2.1.2
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
PyWavelets==1.4.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
threadpoolctl==3.1.0
tifffile==2023.4.12
Werkzeug==2.3.3
zipp==3.15.0
